// packages/shared/src/card-catalog.generated.ts
export const STANDARD_CARD_DEFINITIONS = [
  {
    "cardDefinitionId": "discard",
    "category": "TRADE",
    "element": "PHYSICAL",
    "nameTextKey": "card.discard.name",
    "effectTextKey": "card.discard.effect",
    "instructions": [
      {
        "kind": "SPECIAL",
        "operation": "DISCARD"
      }
    ]
  },
// ...
] as const satisfies readonly CardDefinition[];